Former-commit-id: 19b9941f0bb78520ea13cec794f73b82805d0051

+++ common-tools/clas-detector/src/main/java/org/jlab/detector/decode/CLASDecoder.java
@@ -349,7 +349,7 @@ public class CLASDecoder {
         int    localRun = this.codaDecoder.getRunNumber();
         int  localEvent = this.codaDecoder.getEventNumber();
         long  timeStamp = this.codaDecoder.getTimeStamp();
-        int triggerBits = this.codaDecoder.getTriggerBits();
+        long triggerBits = this.codaDecoder.getTriggerBits();
         
         if(nrun>0){
             localRun = nrun;
@@ -357,7 +357,7 @@ public class CLASDecoder {
         }
         bank.setInt("run",        0, localRun);
         bank.setInt("event",      0, localEvent);
-        bank.setInt("trigger",    0, triggerBits);        
+        bank.setLong("trigger",    0, triggerBits);        
         bank.setFloat("torus",    0, torus);
         bank.setFloat("solenoid", 0, solenoid);        
         bank.setLong("timestamp", 0, timeStamp);
